+++ viewbinding-ktx/src/main/kotlin/ViewBindingDelegate.kt
@@ -1,9 +1,9 @@
 package com.redmadrobot.extensions.viewbinding
 
-import android.os.Handler
-import android.os.Looper
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.Lifecycle.Event.ON_DESTROY
+import androidx.lifecycle.Lifecycle.State.INITIALIZED
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.viewbinding.ViewBinding
@@ -43,7 +43,9 @@ internal class ViewBindingDelegate<VB : ViewBinding> constructor(
 ) : ReadOnlyProperty<Any?, VB>, LifecycleEventObserver {
 
     private var binding: VB? = null
-    private val handler = Handler(Looper.getMainLooper())
+
+    private val fragmentViewNotDestroyed: Boolean
+        get() = fragment.viewLifecycleOwner.lifecycle.currentState.isAtLeast(INITIALIZED)
 
     init {
         fragment.viewLifecycleOwnerLiveData.observe(fragment) {
@@ -51,20 +53,33 @@ internal class ViewBindingDelegate<VB : ViewBinding> constructor(
         }
     }
 
-    override fun getValue(thisRef: Any?, property: KProperty<*>): VB = binding ?: obtainBinding()
+    override fun getValue(thisRef: Any?, property: KProperty<*>): VB {
+        checkFragmentViewNotChanged()
+        return binding ?: obtainBinding()
+    }
+
+    /**
+     * In some cases [Fragment] can change view without triggering [ON_DESTROY] on the old view.
+     * So we should check that bound view is equal to fragment view.
+     */
+    private fun checkFragmentViewNotChanged() {
+        val boundView = binding?.root ?: return
+        if (boundView != fragment.view) binding = null
+    }
 
     private fun obtainBinding(): VB {
         val view = checkNotNull(fragment.view) {
             "ViewBinding is only valid between onCreateView and onDestroyView."
         }
-        return viewBindingClass.bind(view)
-            .also { binding = it }
+        return viewBindingClass.bind(view).also {
+            binding = it.takeIf { fragmentViewNotDestroyed }
+        }
     }
 
     override fun onStateChanged(source: LifecycleOwner, event: Lifecycle.Event) {
-        if (event == Lifecycle.Event.ON_DESTROY) {
+        if (event == ON_DESTROY) {
             source.lifecycle.removeObserver(this)
-            handler.post { binding = null }
+            binding = null
         }
     }
 }