+++ viewbinding-ktx/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Unreleased
 
+## 4.1.2-1 (2021-03-07)
+
+### Fixed
+
+- Add workaround for the case when fragment view changed without triggering ON_DESTROY
+- Save binding only if fragment view is not destroyed
+
 ## 4.1.2-0 (2021-03-01)
 
 ### Dependencies